Add react router, move home out of app.js and make new empty components
