+++ public/index.html
@@ -29,6 +29,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <!-- <img src={profilePicture} className="profile-pic" alt="Jaz's profile" /> -->
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
