+++ src/App.scss
@@ -1,85 +1,1 @@
 @import "./global.scss";
-
-.App-header {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-
-  @media (max-width: $desktop-breakpoint) {
-    flex-direction: column;
-  }
-}
-
-.welcome-content {
-  font-family: $font-stack;
-  display: block;
-  margin-left: 2%;
-  margin-right: 20%;
-  @media (max-width: $desktop-breakpoint) {
-    margin: 2em;
-  }
-}
-.title {
-  font-size: 6vw;
-  color: $font-color;
-  display: flex;
-  margin-bottom: 0.3em;
-  .greeting {
-    font-weight: bold;
-  }
-  @media (max-width: $mobile-breakpoint) {
-    font-size: 10vw;
-  }
-}
-.bio {
-  color: $font-color;
-  line-height: 1.6em;
-  font-size: 16px;
-}
-.menu {
-  display: inline-block;
-  margin-top: 0.1em;
-}
-a.button {
-  display:inline-block;
-  padding:0.5em 1.2em;
-  margin-right: 1em;
-  border-radius: 2em;
-  box-sizing: border-box;
-  text-decoration: none;
-  color: $base-color;
-  font-weight: bold;
-  text-align:center;
-  transition: all 0.2s;
-  @media (max-width: $mobile-breakpoint) {
-    margin-right: 1em;
-    margin-top: 0.5em;
-    padding: 0.5em 0.7em;
-  }
-}
-a.button:hover {
-  opacity: 0.7;
-}
-
-.profile-pic {
-  border-radius: 50%;
-  max-width: 20em;
-  height: 20em;
-  margin-left: 20%;
-  margin-right: 2%;
-  @media (max-width: $desktop-breakpoint) {
-    margin: 3em 3em 0 3em;
-  }
-}
-
-.disabled {
-  opacity: 0.3;
-  pointer-events: none;
-}
-
-//itype
-.ityped-cursor {
-  opacity: 0.5;
-}
