+++ src/components/experience/Experience.js
@@ -2,7 +2,8 @@ import './experience.scss';
 
 function Experience() {
     return (
-    <div className="App">
+    <div>
+        <h1 className="white">Page under construction, please come back later 😊</h1>
     </div>
     );
 }
