+++ src/components/experience/experience.scss
@@ -1,1 +1,2 @@
 @import "src/global.scss";
+@import "src/global.scss";
