+++ src/components/navbar/Navbar.js
@@ -1,6 +1,11 @@
 import './navbar.scss';
 
 function Navbar() {
+
+    constructor(props) {
+        super(props)
+        this.state = { matches: window.matchMedia("(min-width: 768px)").matches };
+    }
     return (
     <div className="App">
     </div>
